Add Organisation model
Initially, we think we want to be able to model organisations as having
many channels, so this commit creates a basic model we can use for this.

At this early stage, we're only adding the name and the slug for the
Organisation as we don't know what else we'll need yet.