Add Channel model
At this early stage, we're only adding the name, slug and decription for
the Channel as we don't know what else we'll need yet.

We weren't sure of the naming here, as Channel is also a keyword in,
I think, ActionCable, so if we're finding this an issue later, know that
there was no deeper thinking behind this and it can be renamed.