Add ChannelsController, routes and show view
We wanted to get something working in a browser at this initial stage,
so this is a very basic controller action, loading the resource and
showing a couple of attribtues in the view.

Note that we haven't added any tests at this initial stage, but as
there's no logic going on at the moment, felt that was ok.