Install tailwind CSS
Tailwind is a ‘utility-first’ CSS framework[1] that we're going to be
trying out in the Sundae Club app.

The installation setup here came from a script via Rails Bytes[2], a
service that provides scripts written by others to help you install
things for your Rails app.

This installs ‘postcss’ too, which I think helps us remove CSS classes
that may be generated, but that we're not actually using. I've also
made a couple of modifications to `tailwind.config.js` to add the
custom font, Sofia Pro.

[1]: https://tailwindcss.com
[2]: https://railsbytes.com/public/templates/z5OsqV