Migration to add suspended_at to Channel model
This was added as part of a livestream in order to have a property on
a Channel that only admins can change. There wasn't any wider thought
put into the feature of 'suspension', so if you're reading this later,
feel free to edit or remove. You can see the livestream at
https://youtu.be/TbRw2jHpgg0