Introduce 'control panel' route namespace with first controller
As part of the stream looking at how we might introduce the concept of
'admin' functionality to the app, this commit adds a controller for
Channels in the 'control panel' namespace. In later commits, we'll
add permissions so that only appropriate users have access to
controllers within this namespace.

Watch the stream where this was added here: https://youtu.be/TbRw2jHpgg0