Add channel suspensions controller in control panel
In this commit we're introducing a new 'suspensions' controller into
the recently added 'controlpanel' namespace, where we're implementing
'admin' functionality.

'Suspension' here was used as example during a stream of something an
admin would want to do, so the feature isn't implemented beyond this.

You can see the stream where this was added here:
https://youtu.be/TbRw2jHpgg0