Add User model
In a previous commit, we installed the Devise gem to prepare for users
to be able to log in to the app. This app creates the `User` model,
which is the resource we'll use with Devise.

The model in this commit was generated by the Devise installation
script, we haven't done any customisation for our app, nor have we added
tests assuming these are all covered by the gem at this early stage.
We'll come back and add feature tests at a later date.

You can see the livestream where we added users at:
https://youtu.be/b-J1TuimRtk