Add 'base' controller within the ControlPanel namespace
This commit was added as part of the livestream looking at adding user
permissions into the app. We're adding a 'base' controller from which
all other controllers in this namespace can inherit in order to add the
`authenticate_user!` before action which will require a user to be
logged in to access any action in the namespace.

Note that this is not checking that the user is the _correct_ user, this
commit would provide access for _any_ logged in user. The work to add
per-user permissions will be added in future commits.

You can see the livestream where we added user permissions at:
https://youtu.be/77Z_u2QDTZA