Install initial pundit policy
Having added our control panel 'base' controller in the previous commit,
we're now adding a method that adds the `control_panel` namespace to all
pundit policies that are called from within the namespace. We could do
this manually for each call to `authorize`, but this saves us some
typing.

We're adding an initial policy to the `ControlPanel::ChannelsController`
here, but it's just a dummy policy checking for the presence of a user.
In future commits, we'll work on ensuring that the correct authorisation
is added here.

You can see the livestream where we added user permissions at:
https://youtu.be/77Z_u2QDTZA