Add Site model
We initially planned on a user creating a Channel and using that as the
primary resource we would use to create a site. As we thought about it,
we realised we may want to add multiple other channels and have other
content on the site. This Site model is intended to represent this
instead.

Here's a link to the livestream where we looked at planning these
changes: https://youtu.be/YDS_DKsCXyo