Add SitesSetup namespace, including Details controller
We're using a pattern here that I've used before for straighforward
step-by-step, or ‘wizard’ user experiences, where we create something
and then use a series of namespaced controllers to take the user through
editing it.

In this commit, we're skipping the update action as we don't yet know
what we want to update and are more interested in getting to the point
where the user will create their account at the end of this setup.