Add users controller to sites setup namespace
In this commit we're adding the controller where we finally create the
user and associate them with the Site they created earlier in the
process. We do have Devise controllers for this, but I don't think we
need them as this does everything we need and Devise model callbacks
should handle any additional requirements e.g. sending confirmation
emails.

Note that there are no tests here, but these are written and will be
added in an upcoming commit.