Fix tests requiring a signed-in confirmed user
This test was broken in earlier work where we added Devise to the base
controller, so we needed to auth the user in the test. After adding a
signed-in user, the test initially failed as the user wasn't
'confirmed'. To resolve this, we're adding a trait to the User factory
and using this to update the failing tests.