Add associations between User and Organisation models
As we were coming up with the initial assocaitions on the livestream,
the association was originally between a User and a Channel. We want to
update that so that these are associated through an Organisation.

Adding the association between a User and an Organisation in this commit
prepares us for that upcoming change.