Use user's organisation for CP channels controller auth
When it was originally written on the livestream, the authorisation
check for the control panel channels controller was "is there a logged
in user". This matched devise's `authenticate_user!` method which we'd
implemented on a previous stream.

In recent work we added an association between User and Organistion.

In this commit we're able to replace our initial and less-than-ideal
"any user" authorisation and ensure that only user that belongs to the
same organisation as the channel can access the control panel channels
controller.