Replace channel/org association with channel/site association
In the initial livestreams, we initially thought an organisation would
have many channels. Later we introduced the idea of the Site model which
would sit between the two so that we can handle sites that had multiple
channels.

This commit removes our initial channel/org association and replaces it
with a channel/site association.