Add initial Github Actions yml file
We are aiming to get rspec tests running in Github when we
make a pull request. This template is a modified version of
one found in this blog post:
https://dev.to/abdellani/github-actions-to-run-rubocop-and-rspec-tests-on-rails-with-postgres-47i

A few important notes:

We're using `ruby/setup-ruby@v1` rather than `actions/setup-ruby@v1` as
this had been deprecated since the blog post was written.

We had to specify a very specific version of Ruby here as
GitHub Actions didn't seem to support our original version.

We're also specifying the version of bundler used to match the one
on my local machine, which is not ideal but we can come back and look
at this separately.