WIP: Add ability to edit videos in Control Panel
In upcoming work, we're going to implement Hotwire (Turbo, Stimulus),
the new technology from Basecamp. In the stream, the approach was "let's
build out a normal version and then we can sprinkle Hotwire on top".

This commit adds a basic editing experience for videos in Control Panel.

Note: No auth in here, so we probably want to add that in when rebasing.